Adding SVM model and fixing up logistic regression model
