+++ addSpaceBefore.py
@@ -1,0 +1,13 @@
+import re
+
+def add_space_before(text):
+    # Add a space before periods, commas, semicolons, etc.
+    processed_text = re.sub(r'([^\s\w])', r' \1', text)
+    return processed_text
+
+# Example usage:
+input_text = """
+Once upon a time in the vibrant city of Harmonyville, there lived a college student named Mia Rodriguez. Mia was a junior majoring in environmental science at Rivertide University. Her love for nature and a deep sense of responsibility towards the planet fueled her determination to make a positive impact. One sunny afternoon, Mia stumbled upon a notice about the annual Green Innovation Challenge—an event where students could pitch eco-friendly projects to a panel of environmental experts. Inspired and eager to contribute, Mia decided to develop a sustainable urban gardening initiative called "GreenHaven." With her hands in the soil and a heart full of passion, Mia transformed an unused corner of the campus into a thriving community garden. She envisioned GreenHaven as a place where students could come together, learn about sustainable agriculture, and cultivate their own fruits and vegetables. Mia believed that this initiative could not only promote environmental consciousness but also foster a sense of community among her peers. As the garden flourished, so did Mia's connection with her fellow students. The project became a hub of creativity, where ideas for sustainable living blossomed alongside the vibrant array of fruits and vegetables. Mia's dedication and leadership drew the attention of both students and faculty alike. When the day of the Green Innovation Challenge arrived, Mia nervously but proudly presented GreenHaven to the panel of judges. The vision, dedication, and positive impact of her project resonated deeply, earning her the first prize and a scholarship for further environmental studies. Word of Mia's success spread, and GreenHaven became a symbol of sustainable living on campus. Mia's journey didn't end with the competition; instead, it marked the beginning of a new chapter. With the scholarship in hand, Mia continued her studies, conducting research on innovative ways to create sustainable urban environments. As Mia graduated from Rivertide University, she left behind a legacy of green initiatives and a campus that had been transformed by the power of community and sustainability. GreenHaven continued to thrive, inspiring future generations of students to think creatively about environmental issues. Mia's story became a beacon of hope, showing that even a single college student with a passion for change could make a lasting impact on the world. And so, as Mia embarked on her journey beyond college, she carried with her not just a degree but the knowledge that small, meaningful actions could ripple into waves of positive transformation for the planet and its people."""
+
+output_text = add_space_before(input_text)
+print(output_text)
